Created routing navigation project
to create a routing project ng new <name> --routing , 
then create your components ng g c <name> -it -is
In index.html verify that <base href="/"> got / 
After, in app-routing.module.ts configures the routes
in @NgModule add RouterModule and don't forget to import it.
routingComponents =>
add it in the app.module.ts :: declarations and imports for routingComponents and AppRoutingModule
To navigates use button and specify the path.